+++ nw/ARENA/TimeRuntimeState.cpp
@@ -1,0 +1,7 @@
+#include <windows.h>
+
+#include "h/super.h"
+#include "storage/h/savefile.h"
+#include "suavik.h"
+
+#include "TimeRuntimeState.inl"