+++ nw/SOUND/LegacyRsxIds.cpp
@@ -1,0 +1,3 @@
+#include <initguid.h>
+
+#include "rsx.h"